contracts/pyproject.toml
[tool.poetry]
name = "contracts"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["John Doe <my@mail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^4.0.0"
python-dotenv = "^1.0.0"
algorand-python = "^3"
algorand-python-testing = "^1"
google-cloud-secret-manager = "^2.26.0"

[tool.poetry.group.dev.dependencies]
algokit-client-generator = "^2.1.0"
puyapy = "*"
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

